pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "assemblyai>=0.43.1",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "logging>=0.4.9.6",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
]
